+++ OS2.1/CPinti/include/IO.h
@@ -25,27 +25,27 @@ namespace cpinti
 	{
 
 		// EXTERNE
-		int 					cpinti_decompress_file			(const char* Source, const char* Destination);
-		int 					cpinti_compress_file			(const char* Source, const char* Destination);
+		long 					cpinti_decompress_file			(const char* Source, const char* Destination);
+		long 					cpinti_compress_file			(const char* Source, const char* Destination);
 		bool 					cpinti_Fichier_Existe			(const char* Source);
-		unsigned int			cpinti_Taille_Fichier			(const char* Source);
-		bool 					cpinti_Lire_Fichier_complet		(const char* Source, const char* Mode, char* _DONNEES, unsigned int TailleFichier);
-		bool 					cpinti_Ecrire_Fichier_complet	(const char* Source, const char* _DONNEES, int Mode);
-		bool 					cpinti_Supprimer_Fichier		(const char* Source, bool Securise, int NombrePasses);
-		bool 					cpinti_Copier_Fichier			(const char* Source, const char* Destination, int Priorite, const char* VAR_Progression, const char* VAR_Octets, const char* VAR_OctetsParSec);
+		unsigned long			cpinti_Taille_Fichier			(const char* Source);
+		bool 					cpinti_Lire_Fichier_complet		(const char* Source, const char* Mode, char* _DONNEES, unsigned long TailleFichier);
+		bool 					cpinti_Ecrire_Fichier_complet	(const char* Source, const char* _DONNEES, long Mode);
+		bool 					cpinti_Supprimer_Fichier		(const char* Source, bool Securise, long NombrePasses);
+		bool 					cpinti_Copier_Fichier			(const char* Source, const char* Destination, long Priorite, const char* VAR_Progression, const char* VAR_Octets, const char* VAR_OctetsParSec);
 	
 		// INTERNE
-		int 					Fichier_decompress		(const char* Source, const char* Destination);
-		int 					Fichier_compress		(const char* Source, const char* Destination);
-		int inf					(FILE *source, FILE *dest);
+		long 					Fichier_decompress		(const char* Source, const char* Destination);
+		long 					Fichier_compress		(const char* Source, const char* Destination);
+		long inf					(FILE *source, FILE *dest);
 		
 		bool 					Fichier_Existe			(const char* Source);
 		bool					Repertoire_Existe		(const char *Source) ;
-		unsigned int 			Taille_Fichier			(const char* Source);
-		bool 					Lire_Fichier_complet	(const char* Source, const char* MODE, char *_DONNEES, unsigned int Taille_Fichier);
-		bool					Ecrire_fichier			(const char* Source, const char* Donnees, int FLAG);
-		bool					Supprimer_Ficher		(const char* Source, bool securise, int NombrePasses);
-		bool 					Copier_Fichier			(const char* Source, const char* Destination, int Priorite, const char* VAR_Progression, const char* VAR_Octets, const char* VAR_OctetsParSec);
+		unsigned long 			Taille_Fichier			(const char* Source);
+		bool 					Lire_Fichier_complet	(const char* Source, const char* MODE, char *_DONNEES, unsigned long Taille_Fichier);
+		bool					Ecrire_fichier			(const char* Source, const char* Donnees, long FLAG);
+		bool					Supprimer_Ficher		(const char* Source, bool securise, long NombrePasses);
+		bool 					Copier_Fichier			(const char* Source, const char* Destination, long Priorite, const char* VAR_Progression, const char* VAR_Octets, const char* VAR_OctetsParSec);
 	}
 }
