+++ OS2.1/CPinti/include/ManagedAlloc.h
@@ -6,8 +6,8 @@
 	{
 		
 		private:
-			unsigned int 	nb_alloc_do			= 0;
-			int 			managed_alloc_max 	= 1024;
+			unsigned long 	nb_alloc_do			= 0;
+			long 			managed_alloc_max 	= 1024;
 			struct alloc_array
 			{
 				void* 			address;
@@ -16,16 +16,16 @@
 			
 			const char* 	name_;
 			
-			unsigned int 	NombreMalloc;
-			unsigned int 	NombreCalloc;
-			unsigned int 	NombreFree;
-			unsigned int 	NombreRealloc;
+			unsigned long 	NombreMalloc;
+			unsigned long 	NombreCalloc;
+			unsigned long 	NombreFree;
+			unsigned long 	NombreRealloc;
 		
 		public:
 
 			void 	ManagedAlloc_			(size_t alloc_max, const char* filename);
 			bool 	ManagedFree				(void* ptr);
-			int 	ManagedAlloc_clean		();
+			long 	ManagedAlloc_clean		();
 			void* 	ManagedMalloc			(size_t size__);
 			void* 	ManagedCalloc			(size_t size__, size_t sizeElem__);
 			void* 	ManagedRealloc			(void* ptr, size_t size__);
