+++ OS2.1/CPinti/include/buffer.h
@@ -2,7 +2,7 @@ namespace cpinti
 {
 
 
-	std::string cpinti_GEST_BUFF				(unsigned int _ID, int _CHEMIN, std::string _DONNEES);
-	int cpinti_GEST_BUFF_c						(unsigned int _ID, int _CHEMIN, const char* _DONNEES);
-	int cpinti_GEST_BUFF_c						(unsigned int _ID, int _CHEMIN, char* _DONNEES);
+	std::string cpinti_GEST_BUFF				(unsigned long _ID, long _CHEMIN, std::string _DONNEES);
+	long cpinti_GEST_BUFF_c						(unsigned long _ID, long _CHEMIN, const char* _DONNEES);
+	long cpinti_GEST_BUFF_c						(unsigned long _ID, long _CHEMIN, char* _DONNEES);
 }
