+++ OS2.1/CPinti/include/int.h
@@ -1,6 +1,6 @@
 
-extern "C" void cpc_SCREEN(int mode);
-extern "C" void cpc_COLOR(int police, int fond);
+extern "C" void cpc_SCREEN(long mode);
+extern "C" void cpc_COLOR(long police, int fond);
 
 namespace cpinti
 {
@@ -9,13 +9,13 @@ namespace cpinti
 		static struct sigaction instance_SigAction[12] = {};
 
 		bool initialise_interception_exception	();
-		void Erreur_Fatale_Signal				(int signalnum);
-		void Erreur_Fatale						(int signalnum, int ligne, char* fichier, char* fonction);
+		void Erreur_Fatale_Signal				(long signalnum);
+		void Erreur_Fatale						(long signalnum, long ligne, char* fichier, char* fonction);
 		
 		void Error_Logo();
 		void Error_Finalise();
 		
-		// void Erreur_Fatale(int signalnum, char const * NomMethode);
+		// void Erreur_Fatale(long signalnum, char const * NomMethode);
 
 	}
 } // namespace cpinti
